Created modal view for program examples
