+++ ARProgramming/ARProgramming/Features/CardDetail/Controller/ExampleProgramViewController.swift
@@ -21,42 +21,10 @@ class ExampleProgramViewController: UIViewController {
             exampleProgramTable.delegate = tableDataSource
         }
     }
-    @IBOutlet weak var previewScene: SCNView! {
-        didSet {
-            previewScene.autoenablesDefaultLighting = true
-            previewScene.scene = SCNScene()
-            previewScene.scene?.rootNode.rotation = SCNVector4(0, -1, 0, 1)
-            previewScene.delegate = self
-            previewScene.isPlaying = true
-            
-            //Load empty level for preview
-            previewLevelViewModel.anchor(at: previewScene.scene?.rootNode)
-            
-            //Set up camera
-            let camera = SCNNode()
-            camera.camera = SCNCamera()
-            camera.camera?.zNear = 0.02
-            camera.camera?.zFar = 10
-            camera.position = SCNVector3(0, 0.9, 0.5)
-            camera.rotation = SCNVector4(-1, -0.033, -0.025, 1.1)
-            previewScene.pointOfView = camera
-        }
-    }
     
     //MARK: - Injected properties
     var tableDataSource: ExampleProgramTableDataSource!
     var levelRepository: LevelRepository!
-    var previewLevelViewModel: LevelViewModeling! {
-        didSet {
-            previewLevelViewModel.display(level: levelRepository.emptylevel)
-            
-            //Add grid floor
-            let ground = SCNNode(geometry: SCNPlane(width: 5, height: 5))
-            ground.eulerAngles.x = -.pi / 2
-            ground.geometry?.materials.first?.diffuse.contents = UIImage(named: "ExampleProgramGridFloor.png")
-            previewLevelViewModel.addNode(ground)
-        }
-    }
     var gameLevelViewModel: LevelViewModeling! {
         didSet {
             gameLevelViewModel.display(level: levelRepository.emptylevel)
@@ -66,9 +34,11 @@ class ExampleProgramViewController: UIViewController {
     func showExamples(forCard card: Card) {
         tableDataSource?.showExamplesForCard(withName: card.internalName)
     }
-    
+
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         switch segue.destination {
+        case let exampleController as ExamplePreviewViewController:
+            exampleController.viewModel = sender as? ProgramsViewModeling
         case let arContainer as ARContainerViewController:
             arContainer.levelViewModel = gameLevelViewModel
         default:
@@ -79,28 +49,7 @@ class ExampleProgramViewController: UIViewController {
 
 //MARK: - ExampleProgramSelectorDelegate
 extension ExampleProgramViewController: ExampleProgramSelectorDelegate {
-    
     func editorSelected(editor: ProgramsViewModeling) {
-        previewLevelViewModel.reset()
-        exampleProgramTable.allowsSelection = false
-        
-        editor.running.onValue = { [weak self] running in
-            self?.exampleProgramTable.allowsSelection = !running
-        }
-        
-        DispatchQueue.main.asyncAfter(wallDeadline: DispatchWallTime.now() + 0.5) { [weak self] in
-            if let entity = self?.previewLevelViewModel.player {
-                editor.start(on: entity)
-            } else {
-                self?.exampleProgramTable.allowsSelection = true
-            }
-        }
-    }
-}
-
-//MARK: - SCNSceneRendererDelegate
-extension ExampleProgramViewController: SCNSceneRendererDelegate {
-    func renderer(_ renderer: SCNSceneRenderer, updateAtTime time: TimeInterval) {
-        previewLevelViewModel.update(currentTime: time)
+        performSegue(withIdentifier: "showProgramExample", sender: editor)
     }
 }
